Move logging into utils folder
